limit public API
